+++ bamboo/create-release-artifacts.sh
@@ -7,7 +7,6 @@ set -ex
 (cd tf-modules/distribution && ./bin/build-tf-module.sh && cp ./dist/terraform-aws-cumulus-distribution.zip ../../terraform-aws-cumulus-distribution.zip)
 (cd tf-modules/s3-replicator && ./bin/build-tf-module.sh && cp ./dist/terraform-aws-cumulus-s3-replicator.zip ../../terraform-aws-cumulus-s3-replicator.zip)
 (cd tf-modules/cumulus-rds-tf && ./bin/build-tf-module.sh && cp ./dist/terraform-aws-cumulus-rds.zip ../../terraform-aws-cumulus-rds.zip)
-(cd lambdas/data-migration1 && ./bin/build-tf-module.sh && cp ./tf-module/terraform-aws-cumulus-data-migrations1.zip ../../terraform-aws-cumulus-data-migrations1.zip)
 
 ## Create zipfiles
 zip -FS -r -x \*node_modules\* \*internal\* -o terraform-aws-cumulus.zip tf-modules lambdas tasks/**/dist/** packages/**/dist/**