+++ example/cumulus-tf/main.tf
@@ -175,7 +175,7 @@ module "cumulus" {
     "mobrien84",
     "nnageswa",
     "npauzenga",
-    "vnguyen",
+    "vnguyen"
   ]
   archive_api_url             = var.archive_api_url
   archive_api_port            = var.archive_api_port