+++ example/cumulus-tf/variables.tf
@@ -362,7 +362,7 @@ variable "rds_admin_access_secret_arn" {
 variable "async_operation_image_version" {
   description = "docker image version to use for Cumulus async operations tasks"
   type = string
-  default = "44"
+  default = "46"
 }
 
 variable "cumulus_process_activity_version" {