+++ example/spec/parallel/testAPI/granuleSpec.js
@@ -153,6 +153,7 @@ describe('The Granules API', () => {
       discoveredGranule = await getGranule({
         prefix,
         granuleId,
+        collectionId,
       });
       expect(discoveredGranule).toEqual(jasmine.objectContaining(randomGranuleRecord));
     });
@@ -185,6 +186,8 @@ describe('The Granules API', () => {
       };
       const response = await updateGranule({
         prefix,
+        granuleId: modifiedGranule.granuleId,
+        collectionId: modifiedGranule.collectionId,
         body: modifiedGranule,
       });
 
@@ -192,6 +195,7 @@ describe('The Granules API', () => {
       updatedGranuleFromApi = await getGranule({
         prefix,
         granuleId: modifiedGranule.granuleId,
+        collectionId: modifiedGranule.collectionId,
       });
       updatedGranuleFromApi.execution = undefined;
       expect(updatedGranuleFromApi).toEqual(jasmine.objectContaining(modifiedGranule));
@@ -216,6 +220,7 @@ describe('The Granules API', () => {
       updatedGranuleFromApi = await getGranule({
         prefix,
         granuleId,
+        collectionId,
       });
       expect(updatedGranuleFromApi.execution).toBe(executionRecord.execution);
     });
@@ -364,6 +369,8 @@ describe('The Granules API', () => {
       try {
         await updateGranule({
           prefix,
+          granuleId: granuleWithDiffCollection.granuleId,
+          collectionId: granuleWithDiffCollection.collectionId,
           body: granuleWithDiffCollection,
         });
       } catch (error) {