+++ packages/api/app/routes.js
@@ -7,7 +7,6 @@ const { defaultErrorHandler } = require('./middleware');
 
 const collections = require('../endpoints/collections');
 const granules = require('../endpoints/granules');
-const granuleCsv = require('../endpoints/granule-csv');
 const providers = require('../endpoints/providers');
 const pdrs = require('../endpoints/pdrs');
 const rules = require('../endpoints/rules');
@@ -25,7 +24,6 @@ const version = require('../endpoints/version');
 const workflows = require('../endpoints/workflows');
 const dashboard = require('../endpoints/dashboard');
 const elasticsearch = require('../endpoints/elasticsearch');
-const migrationCounts = require('../endpoints/migrationCounts');
 const deadLetterArchive = require('../endpoints/dead-letter-archive');
 const { launchpadProtectedAuth } = require('./launchpadAuth');
 const launchpadSaml = require('../endpoints/launchpadSaml');
@@ -43,17 +41,11 @@ if (process.env.FAKE_AUTH === 'true') {
 // dead letters endpoint
 router.use('/deadLetterArchive', ensureAuthorized, deadLetterArchive.router);
 
-//migrationCounts endpoint
-router.use('/migrationCounts', ensureAuthorized, migrationCounts.router);
-
 // collections endpoints
 router.use('/collections', ensureAuthorized, collections.router);
 
 // granules endpoints
 router.use('/granules', ensureAuthorized, granules.router);
-
-// granule csv endpoints
-router.use('/granule-csv', ensureAuthorized, granuleCsv);
 
 // provider endpoints
 router.use('/providers', ensureAuthorized, providers.router);