+++ packages/db/src/test-utils.ts
@@ -4,6 +4,7 @@ import { v4 as uuidv4 } from 'uuid';
 import pRetry from 'p-retry';
 
 import { getExecutionUrlFromArn } from '@cumulus/message/Executions';
+import { randomId } from '@cumulus/common/test-utils';
 
 import { getKnexClient } from './connection';
 import { localStackConnectionEnv } from './config';
@@ -60,7 +61,7 @@ export const fakeRuleRecordFactory = (
   params: Partial<PostgresRule>
 ): PostgresRule => ({
   name: cryptoRandomString({ length: 8 }),
-  workflow: 'Random Workflow',
+  workflow: randomId('workflow'),
   type: 'onetime',
   enabled: false,
   created_at: new Date(),
@@ -109,6 +110,8 @@ export const fakeProviderRecordFactory = (
   name: `provider${cryptoRandomString({ length: 5 })}`,
   host: 'test-bucket',
   protocol: 's3',
+  created_at: new Date(),
+  updated_at: new Date(),
   ...params,
 });