+++ packages/es-client/search.js
@@ -23,6 +23,8 @@ const logDetails = {
 };
 
 const defaultIndexAlias = 'cumulus-alias';
+const multipleRecordFoundString = 'More than one record was found!';
+const recordNotFoundString = 'Record not found';
 
 const getCredentials = () =>
   new Promise((resolve, reject) => aws.config.getCredentials((err) => {
@@ -135,12 +137,6 @@ class BaseSearch {
     this.frm = (page - 1) * this.size;
     this.page = Number.parseInt((params.skip) ? params.skip : page, 10);
     this.index = index || defaultIndexAlias;
-
-    if (this.type === process.env.CollectionsTable) {
-      this.hash = 'collectionName';
-    } else if (this.type === process.env.PdrsTable) {
-      this.hash = 'pdrName';
-    }
   }
 
   _buildSearch() {
@@ -234,10 +230,10 @@ class BaseSearch {
     }).then((response) => response.body);
 
     if (result.hits.total > 1) {
-      return { detail: 'More than one record was found!' };
+      return { detail: multipleRecordFoundString };
     }
     if (result.hits.total === 0) {
-      return { detail: 'Record not found' };
+      return { detail: recordNotFoundString };
     }
 
     const resp = result.hits.hits[0]._source;
@@ -247,76 +243,7 @@ class BaseSearch {
 
   async exists(id, parentId) {
     const response = await this.get(id, parentId);
-    return response.detail !== 'Record not found';
-  }
-
-  async granulesStats(key, value) {
-    const body = {
-      query: {
-        term: {
-          [`${key}.keyword`]: value,
-        },
-      },
-      aggs: {
-        statusCount: {
-          terms: {
-            field: 'status.keyword',
-          },
-        },
-        averageDuration: {
-          avg: {
-            field: 'duration',
-          },
-        },
-        granulesCount: {
-          value_count: {
-            field: 'granuleId.keyword',
-          },
-        },
-      },
-    };
-
-    const ag = await this.client.search({
-      index: this.index,
-      type: process.env.GranulesTable,
-      body: body,
-      size: 0,
-    });
-
-    const status = {
-      failed: 0,
-      ingesting: 0,
-      processing: 0,
-      archiving: 0,
-      cmr: 0,
-      completed: 0,
-    };
-
-    const item = ag.body.aggregations;
-
-    const newObj = {
-      averageDuration: item.averageDuration.value,
-      granules: item.granulesCount.value,
-      granulesStatus: { ...status },
-    };
-
-    item.statusCount.buckets.forEach((b) => {
-      newObj.granulesStatus[b.key] = b.doc_count;
-    });
-
-    if (newObj.granules > 0) {
-      newObj.progress = (
-        (
-          (newObj.granulesStatus.completed + newObj.granulesStatus.failed)
-          / newObj.granules
-        )
-        * 100
-      );
-    } else {
-      newObj.progress = 0;
-    }
-
-    return newObj;
+    return response.detail !== recordNotFoundString;
   }
 
   async query(searchParamsOverride) {
@@ -378,5 +305,7 @@ module.exports = {
   BaseSearch,
   Search,
   defaultIndexAlias,
+  multipleRecordFoundString,
+  recordNotFoundString,
   getLocalEsHost,
 };